+++ CHANGES.rst
@@ -27,6 +27,9 @@ Features
 Other changes
 +++++++++++++
 
+- Make showing the ZMI modal add dialog configurable per product
+  (`#535 <https://github.com/zopefoundation/Zope/issues/535>`_)
+
 - renamed/moved the `Zope 2 Book` to `Zope Book`
   (`#443 <https://github.com/zopefoundation/Zope/issues/443>`_)